+++ HDNXUdemyModel/Model/ContentCourseDetail.cs
@@ -4,7 +4,7 @@ namespace HDNXUdemyModel.Model
 {
     public class ContentCourseDetailModel : BaseModel
     {
-        public int IdContent { get; set; }
+        public long IdContent { get; set; }
 
         public string? NameSubContent { get; set; }
